refactor to addCredential, new CloudConnectionForm component
